requirements.txt
python-dotenv==1.0.0
confluent_kafka==2.3.0
sqlalchemy==1.4.49
pandas==1.3.5
pyodbc==4.0.39
pymssql
google-generativeai
langchain-google-genai 
streamlit